+++ lib/src/utils/vector_math/src/vector_math_geometry/filters/invert_filter.dart
@@ -7,10 +7,6 @@ part of '../../../vector_math_geometry.dart';
 class InvertFilter extends InplaceGeometryFilter {
   @override
   void filterInplace(MeshGeometry mesh) {
-    // TODO: Do the tangents need to be inverted? Maybe just the W component?
-    // TODO: Should modify in-place be allowed, or should it be required
-    // to return a new geometry?
-
     // Swap all the triangle indices
     final indicies = mesh.indices!;